requirements.txt
Flask==2.3.4
mysql-connector-python==8.1.1
flask-cors==3.1.1
